Added validators for api

+++ opinions_app/api_views.py
@@ -6,13 +6,17 @@ from flask import jsonify, request
 from . import app, db
 from .models import Opinion
 from .views import random_opinion
+from .error_handlers import InvalidAPIUsage
 
 
 # Явно разрешить метод GET
 @app.route('/api/opinions/<int:id>/', methods=['GET'])
 def get_opinion(id):
     # Получить объект по id или выбросить ошибку
-    opinion = Opinion.query.get_or_404(id)
+    opinion = Opinion.query.get(id)
+    if opinion is None:
+        # Тут код ответа нужно указать явным образом
+        raise InvalidAPIUsage('Мнение с указанным id не найдено', 404)
     # Конвертировать данные в JSON и вернуть объект и код ответа API
     return jsonify({'opinion': opinion.to_dict()}), 200
 
@@ -20,9 +24,15 @@ def get_opinion(id):
 @app.route('/api/opinions/<int:id>/', methods=['PATCH'])
 def update_opinion(id):
     data = request.get_json()
-    opinion = Opinion.query.get_or_404(id)
-    # Если метод get_or_404 не найдёт указанный ключ,
-    # то он выбросит исключение 404
+    if (
+        'text' in data and 
+        Opinion.query.filter_by(text=data['text']).first() is not None
+    ):
+        raise InvalidAPIUsage('Такое мнение уже есть в базе данных')
+    opinion = Opinion.query.get(id)
+    # Тут код ответа нужно указать явным образом
+    if opinion is None:
+        raise InvalidAPIUsage('Мнение с указанным id не найдено', 404)
     opinion.title = data.get('title', opinion.title)
     opinion.text = data.get('text', opinion.text)
     opinion.source = data.get('source', opinion.source)
@@ -35,7 +45,10 @@ def update_opinion(id):
 
 @app.route('/api/opinions/<int:id>/', methods=['DELETE'])
 def delete_opinion(id):
-    opinion = Opinion.query.get_or_404(id)
+    opinion = Opinion.query.get(id)
+    if opinion is None:
+        # Тут код ответа нужно указать явным образом
+        raise InvalidAPIUsage('Мнение с указанным id не найдено', 404)
     db.session.delete(opinion)
     db.session.commit()
     # При удалении принято возвращать только код ответа 204
@@ -56,6 +69,17 @@ def get_opinions():
 def add_opinion():
     # Получение данные из запроса в виде словаря
     data = request.get_json()
+    # Если нужных ключей нет в словаре,
+    if 'title' not in data or 'text' not in data:
+        # Выбрасываем собственное исключение.
+        # Второй параметр (статус-код) можно не передавать:
+        # нужно вернуть код 400, а именно он возвращается по умолчанию
+        raise InvalidAPIUsage('В запросе отсутствуют обязательные поля')
+    # Если в базе данных уже есть объект
+    # с таким же значением поля text
+    if Opinion.query.filter_by(text=data['text']).first() is not None:
+        # Выбрасываем собственное исключение
+        raise InvalidAPIUsage('Такое мнение уже есть в базе данных')
     # Создание нового пустого экземпляра модели
     opinion = Opinion()
     # Наполнение его данными из запроса
@@ -70,4 +94,7 @@ def add_opinion():
 @app.route('/api/get-random-opinion/', methods=['GET'])
 def get_random_opinion():
     opinion = random_opinion()
-    return jsonify({'opinion': opinion.to_dict()}), 200
+    if opinion is not None:
+        return jsonify({'opinion': opinion.to_dict()}), 200
+    # Тут код ответа нужно указать явным образом
+    raise InvalidAPIUsage('В базе данных нет мнений', 404)
